+++ lib/td_bg_web/controllers/business_concept_version_search_controller.ex
@@ -70,8 +70,7 @@ defmodule TdBgWeb.BusinessConceptVersionSearchController do
        }) do
     results =
       %{results: business_concept_versions} =
-      params
-      |> Search.search_business_concept_versions(claims, page, size)
+      Search.search_business_concept_versions(params, claims, page, size)
 
     bcv_with_links =
       business_concept_versions